requirements.txt
streamlit==1.32.0
pandas==2.1.4
scikit-learn==1.4.0
numpy==1.26.3
google-generativeai==0.3.2
python-dotenv==1.0.1 